fix(#101 WI-2a): Gate-4 fixes — minute-level average rounding; split per-book stats tests
- Medium: average rounds at the MINUTE level (90-119s avg now '2m', not
  '1m'); parameterized boundary tests 89/90/95/119/120/149/150s
- Low: per-book stats suites moved to PersistenceActorPerBookStatsTests
  with shared helpers (read-tests file back under budget)
